+++ NEWS.md
@@ -1,4 +1,6 @@
 # MSTest 0.1.1.9000
+* Made change to MSARmdl(), MSVARmdl(), and HMmdl(). Specifically, when msmu or msvar is FALSE, output list copies single regime value for each k regime. This is needed for simulating the null when either msmu or msvar is FALSE in LMCLRTest() and MMCLRTest().
+* Added option to use different number of initial values for estimating MSMs with observed data vs. for null distribution (see documentation for LMCLRTest() and MMCLRTest()).
 
 # MSTest 0.1.1
 
@@ -6,7 +8,7 @@
 * Added more examples. Examples that take long to complete are commented out but serve to get familiar with usage. 
 * Fixed bug related to using init_theta (setting initial values of parameters) when estimating Markov models (i.e. MSARmdl(), HMmdl(), and MSVARmdl()). 
 * Fixed bug in MMC_bound() when k0>1
-* updated `DESCRIPTION` file for new version.
+* Updated `DESCRIPTION` file for new version.
 
 # MSTest 0.1.0