+++ NEWS.md
@@ -1,6 +1,7 @@
 # MSTest 0.1.1.9000
 * Made change to MSARmdl(), MSVARmdl(), and HMmdl(). Specifically, when msmu or msvar is FALSE, output list copies single regime value for each k regime. This is needed for simulating the null when either msmu or msvar is FALSE in LMCLRTest() and MMCLRTest().
 * Added option to use different number of initial values for estimating MSMs with observed data vs. for null distribution (see documentation for LMCLRTest() and MMCLRTest()).
+* Updated USGNP data set to include 2022 Q3.
 
 # MSTest 0.1.1